requirements.txt
django
ultralytics
pillow
python-dotenv